prefs: add template expantion for schema-id and path
We will need to expand templates for subpages which could have a schema
id or path defined from a mapping key. For example, to modify the settings
for C, you might show "languages.id" page with a mapping of {id} to "c".

When showing the subpage, we will ::connect() to the GSettings post-
resolving the template.