libide: add IdeUri
Ported from lingering GUri prototpyes.

This is going to be used to route "loaders" for the workbench. We need
to cleanup our document/view split, and we can do that if we allow
perspectives/etc to say they can load something. IdeUri is going to be
our data type used to determine the "noun" of what to load.