libide: add IdeLoader interface
I'm not super fond of the name, but the purpose of this interface is
to allow plugins to register loaders for files. A glade loader might
load a file with the Glade designer instead of the code editor. However,
the code editor might also be able to open the file.

We can display this in the "Open With" menu, but also use the priorities
to automatically "do the right thing".