libide: start on tool abstraction
One of the last major parts to get solved is how we will handle tools/
now that libgnome-builder is gone.

To simplify the creation of tools, they can now be implemented as plugins.
Simply define X-Tool-Name: in your plugin definition and implement
IdeApplicationTool. Then the `ide' program can be used to execute your
tool such as "ide build <args>".

We still need to port the tools over, but this lays the framework for that
to happen and continue using IdeApplication.