build: fix in-tree subprocess worker
This lets us spawn gnome-builder-worker from in-tree and both locate
the file and deal with lt-* wrappers.

This is no longer necessary since we ship gnome-builder-worker now.
They are both essentially the same process since all the work happens
in libide, but 20k isn't significantly different from doing a symlink.

I'm not against doing a symlink for this either though if someone knows
the automake trick for that.